Detach Accordion from editorial-components
