+++ .changeset/stupid-rice-collect.md
@@ -1,0 +1,5 @@
+---
+"nextjs-website": patch
+---
+
+Detach Accordion from editorial-components
