+++ apps/nextjs-website/src/app/theme.ts
@@ -11,6 +11,7 @@ export const theme = createTheme(MUIItaliaTheme, {
   palette: {
     custom: {
       backgroundColorDark: '#0B3EE3',
+      bgColorLightGray: '#FAFAFA',
     },
   },
 });
