--- apps/nextjs-website/src/react-components/components/Accordion/index.ts
@@ -1,2 +1,0 @@
-export * from './Accordion';
-export * from './AccordionItem';
