--- apps/nextjs-website/src/react-components/index.ts
@@ -1,3 +1,0 @@
-import { Accordion } from './components/Accordion/Accordion';
-
-export { Accordion };
